+++ packages/core/discovery.config.default.js
@@ -146,11 +146,11 @@ module.exports = {
     enableRedirects: false,
     enableSearchSSR: false,
     enableFaststoreMyAccount: false,
-    enableVtexAssetsLoader: false,
     graphqlCacheControl: {
       maxAge: 0, // 0 disables cache, 5 * 60 enable cache control maxAge 5 minutes
       staleWhileRevalidate: 60 * 60, // 1 hour
     },
     refreshToken: false,
+    scrollRestoration: false,
   },
 }